backend/requirements.txt
asgiref==3.5.0
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
packaging==21.3
Pillow==9.0.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
redis==4.1.3
sqlparse==0.4.2
wrapt==1.13.3
